refactor(Builder): move components to single files and create hooks
