+++ packages/fluentui/react-builder/src/components/Designer.tsx
@@ -1,14 +1,10 @@
 import * as React from 'react';
-import { useImmerReducer, Reducer } from 'use-immer';
 import DocumentTitle from 'react-document-title';
-import { Text, Button, Divider, Accordion } from '@fluentui/react-northstar';
-import { FilesCodeIcon, AcceptIcon, ErrorIcon } from '@fluentui/react-icons-northstar';
+import { Text, Button, Divider } from '@fluentui/react-northstar';
+import { FilesCodeIcon, AcceptIcon } from '@fluentui/react-icons-northstar';
 import { EventListener } from '@fluentui/react-component-event-listener';
 import { renderElementToJSX, CodeSandboxExporter, CodeSandboxState } from '@fluentui/docs-components';
-
 import { componentInfoContext } from '../componentInfo/componentInfoContext';
-import { ComponentInfo } from '../componentInfo/types';
-
 // import Anatomy from './Anatomy';
 import { BrowserWindow } from './BrowserWindow';
 import { Canvas } from './Canvas';
@@ -16,32 +12,18 @@ import { Description } from './Description';
 import { Knobs } from './Knobs';
 import { List } from './List';
 import { Toolbar } from './Toolbar';
-
-import {
-  jsonTreeCloneElement,
-  jsonTreeDeleteElement,
-  jsonTreeFindElement,
-  jsonTreeFindParent,
-  renderJSONTreeToJSXElement,
-  getCodeSandboxInfo,
-  resolveDraggingElement,
-  resolveDrop,
-} from '../config';
-import { readTreeFromStore, readTreeFromURL, writeTreeToStore, writeTreeToURL } from '../utils/treeStore';
-
-import { DesignerMode, JSONTreeElement } from './types';
+import { jsonTreeFindElement, renderJSONTreeToJSXElement, getCodeSandboxInfo, resolveDraggingElement } from '../config';
+import { writeTreeToStore, writeTreeToURL } from '../utils/treeStore';
+import { JSONTreeElement } from './types';
 import { ComponentTree } from './ComponentTree';
 import { GetShareableLink } from './GetShareableLink';
 import { ErrorBoundary } from './ErrorBoundary';
 import { InsertComponent } from './InsertComponent';
-
-import * as axeCore from 'axe-core';
+import { debug, useDesignerState } from '../state';
+import { useAxeOnElement, useMode } from '../hooks';
+import { ErrorPanel } from './ErrorPanel';
 
 const HEADER_HEIGHT = '3rem';
-
-function debug(...args) {
-  console.log('--Designer', ...args);
-}
 
 const CodeEditor = React.lazy(async () => {
   const _CodeEditor = (await import(/* webpackChunkName: "codeeditor" */ './CodeEditor')).CodeEditor;
@@ -49,280 +31,6 @@ const CodeEditor = React.lazy(async () => {
     default: _CodeEditor,
   };
 });
-
-function getDefaultJSONTree(): JSONTreeElement {
-  return { uuid: 'builder-root', type: 'div' };
-}
-
-const focusTreeTitle = uuid => {
-  // TODO: use refs
-  const element = document.querySelector(`#${uuid} [data-is-focusable]`) as HTMLElement;
-  element && element.focus();
-};
-
-type JSONTreeOrigin = 'store' | 'url';
-
-type DesignerState = {
-  draggingElement: JSONTreeElement;
-  jsonTree: JSONTreeElement;
-  jsonTreeOrigin: JSONTreeOrigin;
-  selectedComponentInfo: ComponentInfo; // FIXME: should be computed in render?
-  selectedJSONTreeElementUuid: JSONTreeElement['uuid'];
-  showCode: boolean;
-  code: string | null; // only valid if showCode is set to true
-  codeError: string | null;
-  history: Array<JSONTreeElement>;
-  redo: Array<JSONTreeElement>;
-  insertComponent: { uuid: string; where: string; parentUuid?: string };
-};
-
-type DesignerAction =
-  | { type: 'DRAG_START'; component: JSONTreeElement }
-  | { type: 'DRAG_ABORT' }
-  | { type: 'DRAG_DROP'; dropParent: JSONTreeElement; dropIndex: number }
-  | { type: 'DRAG_CLONE' }
-  | { type: 'DRAG_MOVE' }
-  | { type: 'SELECT_COMPONENT'; component: JSONTreeElement }
-  | { type: 'SELECT_PARENT' }
-  | { type: 'DELETE_SELECTED_COMPONENT' }
-  | { type: 'PROP_CHANGE'; component: JSONTreeElement; propName: string; propValue: any }
-  | { type: 'PROP_DELETE'; component: JSONTreeElement; propName: string }
-  | { type: 'SWITCH_TO_STORE' }
-  | { type: 'RESET_STORE' }
-  | { type: 'SHOW_CODE'; show: boolean }
-  | { type: 'SOURCE_CODE_CHANGE'; code: string; jsonTree: JSONTreeElement }
-  | { type: 'SOURCE_CODE_ERROR'; code: string; error: string }
-  | { type: 'UNDO' }
-  | { type: 'REDO' }
-  | { type: 'OPEN_ADD_DIALOG'; uuid: string; where: string; parent?: string }
-  | { type: 'CLOSE_ADD_DIALOG' }
-  | { type: 'ADD_COMPONENT'; component: string; module: string };
-
-const stateReducer: Reducer<DesignerState, DesignerAction> = (draftState, action) => {
-  debug(`stateReducer: ${action.type}`, { action, draftState: JSON.parse(JSON.stringify(draftState)) });
-  let treeChanged = false;
-
-  switch (action.type) {
-    case 'DRAG_START':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      draftState.draggingElement = action.component;
-      break;
-
-    case 'DRAG_ABORT':
-      draftState.history.pop();
-
-      draftState.draggingElement = null;
-      break;
-
-    case 'DRAG_DROP':
-      if (action.dropParent) {
-        const dropParent = jsonTreeFindElement(draftState.jsonTree, action.dropParent.uuid);
-        resolveDrop(draftState.draggingElement, dropParent, action.dropIndex);
-        treeChanged = true;
-      }
-
-      const addedComponent = jsonTreeFindElement(draftState.jsonTree, draftState.draggingElement.uuid);
-
-      draftState.draggingElement = null;
-      if (addedComponent) {
-        draftState.selectedJSONTreeElementUuid = addedComponent.uuid;
-        draftState.selectedComponentInfo = componentInfoContext.byDisplayName[addedComponent.displayName];
-      }
-      break;
-
-    case 'DRAG_CLONE':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      draftState.draggingElement = jsonTreeCloneElement(
-        draftState.jsonTree,
-        jsonTreeFindElement(draftState.jsonTree, draftState.selectedJSONTreeElementUuid),
-      );
-      break;
-
-    case 'DRAG_MOVE':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      draftState.draggingElement = jsonTreeCloneElement(
-        draftState.jsonTree,
-        jsonTreeFindElement(draftState.jsonTree, draftState.selectedJSONTreeElementUuid),
-      );
-      jsonTreeDeleteElement(draftState.jsonTree, draftState.selectedJSONTreeElementUuid);
-      treeChanged = true;
-      break;
-
-    case 'SELECT_COMPONENT':
-      if (action.component && draftState.selectedJSONTreeElementUuid !== action.component.uuid) {
-        draftState.selectedJSONTreeElementUuid = action.component.uuid;
-        draftState.selectedComponentInfo = componentInfoContext.byDisplayName[action.component.displayName];
-      } else {
-        draftState.selectedJSONTreeElementUuid = null;
-        draftState.selectedComponentInfo = null;
-      }
-      break;
-
-    case 'SELECT_PARENT':
-      const parent = jsonTreeFindParent(draftState.jsonTree, draftState.selectedJSONTreeElementUuid);
-      if (parent) {
-        draftState.selectedJSONTreeElementUuid = parent.uuid;
-        draftState.selectedComponentInfo = componentInfoContext.byDisplayName[parent.displayName];
-      }
-      break;
-
-    case 'DELETE_SELECTED_COMPONENT':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      if (draftState.selectedJSONTreeElementUuid) {
-        jsonTreeDeleteElement(draftState.jsonTree, draftState.selectedJSONTreeElementUuid);
-        draftState.selectedJSONTreeElementUuid = null;
-        draftState.selectedComponentInfo = null;
-        treeChanged = true;
-      }
-      break;
-
-    case 'PROP_CHANGE':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      const editedComponent = jsonTreeFindElement(draftState.jsonTree, action.component.uuid);
-      if (editedComponent) {
-        if (!editedComponent.props) {
-          editedComponent.props = {};
-        }
-        editedComponent.props[action.propName] = action.propValue;
-        treeChanged = true;
-      }
-      break;
-
-    case 'PROP_DELETE':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      const component = jsonTreeFindElement(draftState.jsonTree, action.component.uuid);
-      if (component) {
-        if (!component.props) {
-          component.props = {};
-        }
-        delete component.props[action.propName];
-        treeChanged = true;
-      }
-      break;
-
-    case 'SWITCH_TO_STORE':
-      draftState.jsonTree = readTreeFromStore() || getDefaultJSONTree();
-      draftState.jsonTreeOrigin = 'store';
-      treeChanged = true;
-      break;
-
-    case 'RESET_STORE':
-      draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-      draftState.redo = [];
-
-      draftState.jsonTree = getDefaultJSONTree();
-      draftState.jsonTreeOrigin = 'store';
-      treeChanged = true;
-      break;
-
-    case 'SHOW_CODE':
-      try {
-        draftState.showCode = action.show;
-        draftState.code = action.show ? renderElementToJSX(renderJSONTreeToJSXElement(draftState.jsonTree)) : null;
-      } catch (e) {
-        console.error('Failed to convert tree to code.', e.toString());
-      }
-      break;
-
-    case 'SOURCE_CODE_CHANGE':
-      draftState.code = action.code;
-      draftState.selectedJSONTreeElementUuid = null;
-      draftState.selectedComponentInfo = null;
-      draftState.jsonTree = action.jsonTree;
-      draftState.codeError = null;
-      break;
-
-    case 'SOURCE_CODE_ERROR':
-      draftState.code = action.code;
-      draftState.selectedJSONTreeElementUuid = null;
-      draftState.selectedComponentInfo = null;
-      draftState.codeError = action.error;
-      break;
-
-    case 'UNDO':
-      if (draftState.history.length > 0) {
-        draftState.redo.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-        draftState.jsonTree = draftState.history.pop();
-      }
-      break;
-
-    case 'REDO':
-      if (draftState.redo.length > 0) {
-        draftState.history.push(JSON.parse(JSON.stringify(draftState.jsonTree)));
-        draftState.jsonTree = draftState.redo.pop();
-      }
-      break;
-
-    case 'OPEN_ADD_DIALOG': {
-      const parent = jsonTreeFindParent(draftState.jsonTree, action.uuid);
-      draftState.insertComponent = { where: action.where, uuid: action.uuid, parentUuid: `${parent?.uuid}` };
-      break;
-    }
-
-    case 'CLOSE_ADD_DIALOG':
-      draftState.insertComponent = null;
-      break;
-
-    case 'ADD_COMPONENT': {
-      const element = resolveDraggingElement(action.component, action.module);
-
-      let parent: JSONTreeElement = undefined;
-      let index = 0;
-      const { where, uuid, parentUuid } = draftState.insertComponent;
-      draftState.insertComponent = null;
-
-      if (where === 'first') {
-        parent = draftState.jsonTree;
-      } else if (where === 'child') {
-        parent = jsonTreeFindElement(draftState.jsonTree, uuid);
-      } else {
-        parent = jsonTreeFindElement(draftState.jsonTree, parentUuid);
-        index = parent.props.children.findIndex(c => c['uuid'] === uuid);
-        if (index === -1) {
-          index = 0;
-        } else {
-          where === 'after' && index++;
-        }
-      }
-
-      resolveDrop(element, parent, index);
-
-      draftState.selectedJSONTreeElementUuid = element.uuid;
-      draftState.selectedComponentInfo = componentInfoContext.byDisplayName[element.displayName];
-      treeChanged = true;
-      setTimeout(() => focusTreeTitle(element.uuid));
-      break;
-    }
-
-    default:
-      throw new Error(`Invalid action ${action}`);
-  }
-
-  if (treeChanged && draftState.showCode) {
-    draftState.code = renderElementToJSX(renderJSONTreeToJSXElement(draftState.jsonTree));
-    draftState.codeError = null;
-  }
-};
-
-function useMode(): [{ mode: DesignerMode; isExpanding: boolean; isSelecting: boolean }, (mode: DesignerMode) => void] {
-  const [mode, setMode] = React.useState<DesignerMode>('build');
-  const isExpanding = mode === 'build';
-  const isSelecting = mode === 'build' || mode === 'design';
-
-  return [{ mode, isExpanding, isSelecting }, setMode];
-}
 
 export const Designer: React.FunctionComponent = () => {
   debug('render');
@@ -335,67 +43,12 @@ export const Designer: React.FunctionComponent = () => {
 
   const draggingElementRef = React.useRef<HTMLDivElement>();
 
-  const [state, dispatch] = useImmerReducer(stateReducer, null, () => {
-    let jsonTreeOrigin: JSONTreeOrigin = 'url';
-    let jsonTree = readTreeFromURL(window.location.href);
-    if (!jsonTree) {
-      jsonTree = readTreeFromStore() || getDefaultJSONTree();
-      jsonTreeOrigin = 'store';
-    }
-
-    return {
-      draggingElement: null,
-      jsonTree,
-      jsonTreeOrigin,
-      selectedComponentInfo: null,
-      selectedJSONTreeElementUuid: null,
-      showCode: false,
-      code: null,
-      codeError: null,
-      history: [],
-      redo: [],
-      insertComponent: null,
-    };
-  });
-
+  const [state, dispatch] = useDesignerState();
   const [{ mode, isExpanding, isSelecting }, setMode] = useMode();
   const [showJSONTree, handleShowJSONTreeChange] = React.useState(false);
-  const [axeErrors, setAxeErrors] = React.useState([]);
   const [headerMessage, setHeaderMessage] = React.useState('');
 
-  const runAxeOnElement = React.useCallback(selectedElementUuid => {
-    const iframe = document.getElementsByTagName('iframe')[0];
-    const selectedComponentAxeErrors = [];
-    axeCore.run(
-      iframe,
-      {
-        rules: {
-          // excluding rules which are related to the whole page not to components
-          'page-has-heading-one': { enabled: false },
-          region: { enabled: false },
-          'landmark-one-main': { enabled: false },
-        },
-      },
-      (err, result) => {
-        if (err) {
-          console.error('Axe failed', err);
-        } else {
-          result.violations.forEach(violation => {
-            violation.nodes.forEach(node => {
-              if (node.html.includes(`data-builder-id="${selectedElementUuid}"`)) {
-                selectedComponentAxeErrors.push(
-                  node.failureSummary
-                    .replace('Fix all of the following:', '-')
-                    .replace('Fix any of the following:', '-'),
-                );
-              }
-            });
-          });
-        }
-        setAxeErrors(selectedComponentAxeErrors);
-      },
-    );
-  }, []);
+  const [axeErrors, runAxeOnElement] = useAxeOnElement();
 
   React.useEffect(() => {
     if (state.selectedJSONTreeElementUuid) {
@@ -912,40 +565,6 @@ export const Designer: React.FunctionComponent = () => {
           </div>
         )}
       </div>
-    </div>
-  );
-};
-
-const ErrorPanel = ({ axeErrors }) => {
-  const panels = [
-    {
-      key: 'axe',
-      title: {
-        'aria-level': 4,
-        content: (
-          <Text>
-            <ErrorIcon style={{ marginRight: '0.5rem' }} /> {axeErrors.length} Accessibility{' '}
-            {axeErrors.length > 1 ? 'Errors' : 'Error'}
-          </Text>
-        ),
-      },
-      content: (
-        <ul style={{ padding: '0rem 0.7rem', listStyleType: 'none' }}>
-          {axeErrors.map(error => (
-            <li>{error}</li>
-          ))}
-        </ul>
-      ),
-    },
-  ];
-
-  return (
-    <div
-      style={{
-        background: '#e3404022',
-      }}
-    >
-      <Accordion panels={panels} />
     </div>
   );
 };
